third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-release-noreference.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeySession lifetime after release() without references</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            // Since MediaKeySession (but not MediaKeys) are ActiveDOMObjects,
            // we can determine when they are garbage collected.
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)
            async_test(function(test)
            {
                var initDataType;
                var initData;
                var startingActiveDOMObjectCount = window.internals.activeDOMObjectCount(document);

                function numActiveDOMObjectsCreated()
                {
                    return window.internals.activeDOMObjectCount(document) - startingActiveDOMObjectCount;
                }

                // Create 2 sessions.
                var mediaKeys;
                var mediaKeySession1;
                var mediaKeySession2;
                getSupportedInitDataType().then(function(type) {
                    initDataType = type;
                    initData = getInitData(initDataType);
                    return navigator.requestMediaKeySystemAccess('org.w3.clearkey', [{}]);
                }).then(function(access) {
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;

                    // Verify MediaKeys are not an ActiveDOMObject.
                    // In non-Oilpan, numActiveDOMObjectsCreate() == 0.
                    // In Oilpan, numActiveDOMObjectsCreate() <= 2.
                    // (1 MediaKeysInitializer and
                    //  1 MediaKeySystemAccessInitializer).
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 2, 'MediaKeys.create()');

                    mediaKeySession1 = mediaKeys.createSession();
                    return mediaKeySession1.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession1.sessionId && mediaKeySession1.sessionId.length > 0);
                    // Should be 1 MediaKeySession.
                    // In non-Oilpan, numActiveDOMObjectsCreate() == 1.
                    // In Oilpan, numActiveDOMObjectsCreate() <= 4.
                    // (1 MediaKeysInitializer,
                    //  1 MediaKeySystemAccessInitializer,
                    //  1 ContentDecryptionModuleResultPromise and
                    //  1 MediaKeySession).
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 4, 'MediaKeys.createSession(1)');
                    mediaKeySession2 = mediaKeys.createSession();
                    return mediaKeySession2.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession2.sessionId && mediaKeySession2.sessionId.length > 0);
                    // Should be 2 MediaKeySessions.
                    // In non-Oilpan, numActiveDOMObjectsCreate() == 2.
                    // In Oilpan, numActiveDOMObjectsCreate() <= 6.
                    // (1 MediaKeysInitializer,
                    //  1 MediaKeySystemAccessInitializer,
                    //  2 ContentDecryptionModuleResultPromise and
                    //  2 MediaKeySession).
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 6, 'mediaKeys.createSession(2)');
                }).then(function(result) {
                    // Run gc(). All sessions should remain as we have a
                    // reference to each one.
                    return createGCPromise();
                }).then(function(result) {
                    // Should be just the 2 MediaKeySessions.
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 3, 'After gc()');

                    // Close the sessions. Once the close() event is received,
                    // they should get garbage collected as there are no JS
                    // references to them.
                    var promise = mediaKeySession1.close();
                    mediaKeySession1 = null;
                    return promise;
                }).then(function(result) {
                    // Give time so that the close event can be processed by
                    // MediaKeySession.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    // Should be just mediaKeySession2.
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 2);
                    var promise = mediaKeySession2.close();
                    mediaKeySession2 = null;
                    return promise;
                }).then(function(result) {
                    // Provide time for the mediaKeySession2 close event to be
                    // handled.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    assert_less_than_equal(numActiveDOMObjectsCreated(), 1);
                    assert_not_equals(mediaKeys, null);
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'MediaKeySession lifetime after release() without references');
        </script>
    </body>
</html>
